Introduce a ChangeSet screen
This will basically encapsulate all that we have built so far. It is the
main screen for a selected PR.
Also introduced a FileListView to show a list of files. Each item in
this is selectable. For now, upon clicking any item, we show the
FileView corresponding to that file