Introduce a UniversalKeyboardHandler and kb shortcut enabled components
We have our own Vimium!
* A trie is used to 'register' keyboard shortcuts as well as the handler
  that should be invoked upon a full or a partial match. This is a
  common component
* This is leveraged by the UniversalKeyboardHandler - the primary js
  entity used to enable keyboard shortcuts. This registers a window
  listener and takes care of registering and unregistering components in
  the trie
* 2 convenience components - KeyboardEnabledComponent and
  KeyboardEnabledList are also provided. These are the way to create
  components in React side of the js code

UniversalKeyboardHandler is a kotlin 'object' that maintains it's own
state and it is not a React component. Initially, I tried making it a
functional component but that was proving too cumbersome. For it to work
seamlessly, the 'functional' component would have to end up being the
'parent' component for the entire screen and we would have had to
somehow keep passing methods to register shortcuts etc.
Leveraging the power of Kotlin seemed so much simpler, easier to
maintain and handle

TODO:
* Provide a way to unregister a shortcut in the trie
* Write test cases for UniversalKeyboardHandler