Add the functionality of marking a file 'done'
Something doesn't feel right about the way all this is being implemented
right now. As of now, the 'owner' of the command is FileListView.
Somehow, I feel the owner of the command should be the
ChangeSetReviewScreen. I'll read some docs/tutorials to figure if this
a case of needing 'lift the state up' to a higher level.

Also, unrelated to the functional change, the KeyboardEnabledComponent
is updated to receive constructor props. They're not being used, but I
was always looking for a way to do this - and I'll just leave the
changes done so that I have a handy example to refer to