Add comment handler in the entire UI chain
Introduced the notion of a FileCommentHandler to store new comments.
The handler is passed from the top most ChangeSetReviewScreen all the
way to DiffView. The reason it doesn't go all the way to CodeView is
because we want to display the comment box for a given editor over the
sister editor (eg: leftEditor comment box will be shown in the right
editor).
We kept newComments separately since
* Our backend API is a dumb API that doesn't do comment de-dupe (yet).
  So it will be easier to POST comments to the backend if we keep only
  new ones.
* On the UI front, we'd eventually want to give different controls (like
  remove/edit) on newly added comments. Having a different list for the
  same makes things easier

Further, the plan is to make fileView a little less logic-heavy and push
all the rendering logic to DiffView. So, eventually, diffView will be
charged with showing comments for all files - even newly added or
deleted files