Introduce a Git Repo searcher
This nifty utility scans through the user's home directory and hunts
down existing git repos. These are then stored in an in-memory cache and
can be used to automagically match a remote repo to it's location in the
user's home directory. This way, the user doesn't need to spend _any_
effort to get going!