Introduce material-ui-styles to ensure styles load correctly
We are building and injecting css for various components via kotlin css.
However, material UI also seemingly injects it's own styles in the <head> element of the dom.
This seemingly happens whenever a component is loaded for the first time
(though not sure) and by default this css is injected at the end of the
<head> tag - thereby always giving it higher priority.
Thus, the top level render should now always be under a StyleProvider
component which will ensure that all css injected by material ui is
injected at the top of the head tag (that's what even ace does btw so
seems it's a good sensible default to have for external components).
Also ensured that custom css generated by kotlin is injected "beforeend"
of the head tag