Add Github Oauth login
This is how the flow works

Non-Logged-In User:
MainScreen -> Login Component -> initiateGithubLogin
initiateGithubLogin : First creates a token in DB then redirects to
github.
Once redirected back from github, the flow is similar as follows
MainScreen -> Login Component -> loginToGithub.
loginToGithub: This compares the 'state' and makes the final call to
github to retrieve the authToken . This is stored in DB and returned.

Logged-In User:
Main Screen -> Login Component -> githubAccessToken
When githubAccessToken doesn't yield an error, we assume user is logged
in.

This is fairly basic and a lot needs to be worked on
(a) Validation if the auth token is still working (should be done in
githubAccessToken)
(b) Seemingly Suboptimal way to implement it in UI with MainScreen ->
LoginComponent dance
(c) No encryption
(d) No tests for OAuthApi backend . Even the endpoints are poorly
defined

Other implementation notes:
* Created a backend API + DB tables to store oauth credentials
* Removed GettingStarted screen and moved loginComponent to MainScreen
  itself. Always thought GettingStarted was an unnecessary abstraction
  and did away with that