Add functionality to jump to the first hunk in a diff view
This took a lot of tinkering, till I finally saw this:
https://stackoverflow.com/questions/23748743/ace-editor-go-to-line
Came up with a rather ill designed way to handle the state variable that
stores the position of the current hunk index.
Eventually, I think we'll go away from this rigid state variable and use
the cursor position instead.

Also took this opportunity to separate out functinality stuffed within
AceCommentsWrapper that had nothing to do with comments!