Introduce our very own fast3r dark theme for ace editor
Right now, the theme is sort of split half between theme-fast3r_dark.js
and Styles.kt; The idea is to keep moving things out of js and into
kotlin (I was just lazy to do it all).
We have more colors to use - but it seems that either the kotlin mode in
the ace editor OR the editor itself is not in the mood to support it.
Will figure it out in a day or two