Make the action bar a floating bar
To save space, the design recommended to make the action bar a floating
thing overlapping the editor. This can be dismissed since most users
would be familiar with the actions anyway (that capability still needs
to be added. Basically it'll set the display of the actionbar to 'none')

This change involved moving the action bar from review screen to
diffview to code view. This kinda makes perfect sense now and definitely
much better than the previous design where diffView used to somehow add
more actions to the actionbar.

Initially, I tried rendering the actionbar within the main 'div' of the
codeView. But somehow ace is hiding it (or doing some magic anyway).
So the actionbar now belongs to the parent. This may lead to some
positioning trouble but we'll live with that for now