Solves a weird bug of keyboard shortcut highlights not being shown
This is quite strange. When I ran the 'productionWebpack' gradle task,
the resultant javascript was not highlighting partial keyboard shortcuts
(the shortcuts were evidently working though)
There was no error whatsoever!

At first I thought it had something to do with DCE, but that doesn't
seem to be the case either.

So I changed the build task to run the developmentWebpack instead.
What's funny is that the DCE + productionWebpack tasks are still
running. Post that, we run developmentWebpack task. And somehow, the
resultant JS is working as expected.

Strange