Introduce a DragNDropEnabledFileList component
This re-uses the same fileItem that was being used in FileListView.
It made sense to carve out fileItem into it's own file now.

It is interesting to note how the draggable is being implemented.
The fileItem component has an optional prop - DraggableProvided. If this
is provided, then we iterate over certain fields of the
DraggableProvided object and _merge_ them to the attributes of the
ListItem.