Introduce basic automation for electron packaging
Created a top level folder to house everything we need for electron
packaging.
Electron & the packager we use is basically a node module - so it made
sense to house it in an independent folder instead of keeping it under
kotlin.
Also introduced a new task called 'electronPackage' that will copy
compiled artifacts that result from 'buildAllArtifacts' into the
designate electron directory.