Switch to 2 letter shortcuts whenever we generate them
Deleted the previous method that generated either single or double
character shortcuts - based on how many were needed.
Since in the design, we are ubiquitously leaving space for 2 letter
shortcuts, it seemed sensible to have uniformity across the board.

While running the code after making the change, I also realised that the
shortcuts being generated within PullRequestList were not done the right
way - so fixed that too.

One more (unfixed) issue - when we move from ChangesetOverView ->
ChangeSetReview, we generate new shortcuts in the constructor. The
trouble with this is that we do this before the previous ones are
unmounted. Since there is a 'preferred sequence' with which shortcuts
are now generated, we lose out on getting the 'good' shortcuts (starting
with a/d) now. This may automatically get fixed in the new design - but
something to always keep in mind going forward