Implement a quick-fix hack for the issue of odd highlights in DiffView
Basically, a lot of the issues were occuring because we were not using
the _view position_ in LineDiffProcessor. For now, a brute force search
is being used. Need to think deeper as to whether backend should not be
sending any 'edits' with filePosition at all.

Switched the jsDiff algo to use diffWordsWithSpace.
However, even this is not working as well as the one I can see in
vimDiff. So need to see if I should send that diff from the backend as
well

Finally, I've disabled a requirement check in LineDiffProcessor that
helped uncover a lot of issues. I'll bring that one back next week when
I work on refining the diffview