Revamp the comments section of the ChangesetOverviewScreen
This is basically the central component that renders the review comment
list (hard coded for now - to test out the design). I've tried to stick
to the flex-box as much as possible here.

Note: Not tested on resolutions < 4k yet. That's because we still have 1
component - the file list - to go.