Enable numeric shortcut handlers
This enables numeric shortcuts of the form <n>-<x> where
n: Is a number
x: Is a character.

A handler is registered for character 'x'. When a user presses <n>-<x>,
the handler gets invoked with the param 'n'. This enables the handler to
act appropriately.
This works completely independent of the trie. It is intentionally kept
simple by allowing only single character ('x') handlers.
So stuff like 111c or 111g is possible; but 111gf is not.

Also, since it's independent of the trie, we can have 'c-x' as a trie
shortcut and still allow user to press '111c' triggering the numeric
shortcut for 'c'.

Present use cases include allowing for comments on line numbers (111c)
or jumping to a particular line (111g)