Add support for keyboard based comment handling in DiffView
We register 'n'-c as the key for comment based shortcuts.
This diff choses the correct editor based on which one is in focus at
the time (we need another shortcut to switch editors but that's for
another time)

The diff has a bit of a refactoring involved because we were needlessly
using a RowColObject to determine the position of the commentbox whereas
what we really needed was just the row number.

Also, fixed a bug in numeric shortcut handling in KeyboardHandler where
we were not firing kbEvent.preventDefault()