Fix keyboard chip highlight which broke as a sideffect of production js
As has been noted in one of the previous commits (long time ago, from a
galaxy far away). The production js build was a little odd - and it
broke keyboard highlight functionality.
This commit introduces a new style and keeps it at the very end of our
styles list - thereby ensuring that it always gets precedence.
This is a stop gap - we should actually chose between a background style
based on a prop. Will fix it in another diff