Introduce the Help box component
This renders a help section and can be placed on any screen.
Ideally this should be handcrafted for each scren, but for now we have a
simple abstraction that can render a list of sections with each section
having a list of items. One such type of item is a
keyboardShortcutExplainer used to give help about a particular keyboard
shortcut combination