Add real t shirt size to the PR Summary card on the UI
Now, we have the connundrum of needing to call the 'diff' api to get the
actual diff (and hence, a real estimate of the t shirt size for the PR).

However, the way we have presently modelled our APIs, we call the
/review API once again when we select any one of the PRs from the
overview screen.

So, in essence, we end up calling the `git diff` equivalent twice.
This needs to improve somehow.